requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
itsdangerous==2.1.2
Jinja2==3.1.2
mariadb==1.1.6
MarkupSafe==2.1.2
packaging==23.0
python-dotenv==1.0.0
tabulate==0.9.0
Werkzeug==2.2.2
